# api-docs.json
{"openapi":"3.0.3","info":{"title":"k12 API","description":"k12 API documentation","contact":{},"license":{"name":"unlicensed"},"version":"0.0.1"},"servers":[{"url":"http://localhost:8080","description":"Inferred Url"}],"tags":[{"name":"account-resource","description":"Account Resource"},{"name":"public-user-resource","description":"Public User Resource"},{"name":"user-resource","description":"User Resource"}],"paths":{"/api/account":{"get":{"tags":["account-resource"],"summary":"getAccount","operationId":"getAccountUsingGET","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AdminUserDTO"}}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"post":{"tags":["account-resource"],"summary":"saveAccount","operationId":"saveAccountUsingPOST","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdminUserDTO"}}}},"responses":{"200":{"description":"OK"},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/account/change-password":{"post":{"tags":["account-resource"],"summary":"changePassword","operationId":"changePasswordUsingPOST","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PasswordChangeDTO"}}}},"responses":{"200":{"description":"OK"},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/account/reset-password/finish":{"post":{"tags":["account-resource"],"summary":"finishPasswordReset","operationId":"finishPasswordResetUsingPOST","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/KeyAndPasswordVM"}}}},"responses":{"200":{"description":"OK"},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/account/reset-password/init":{"post":{"tags":["account-resource"],"summary":"requestPasswordReset","operationId":"requestPasswordResetUsingPOST","requestBody":{"content":{"application/json":{"schema":{"type":"string"}}}},"responses":{"200":{"description":"OK"},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/account/sessions":{"get":{"tags":["account-resource"],"summary":"getCurrentSessions","operationId":"getCurrentSessionsUsingGET","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PersistentToken"}}}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/account/sessions/{series}":{"delete":{"tags":["account-resource"],"summary":"invalidateSession","operationId":"invalidateSessionUsingDELETE","parameters":[{"name":"series","in":"path","description":"series","required":true,"style":"simple","schema":{"type":"string"}}],"responses":{"200":{"description":"OK"},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}}},"/api/activate":{"get":{"tags":["account-resource"],"summary":"activateAccount","operationId":"activateAccountUsingGET","parameters":[{"name":"key","in":"query","description":"key","required":true,"style":"form","schema":{"type":"string"}}],"responses":{"200":{"description":"OK"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/authenticate":{"get":{"tags":["account-resource"],"summary":"isAuthenticated","operationId":"isAuthenticatedUsingGET","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/register":{"post":{"tags":["account-resource"],"summary":"registerAccount","operationId":"registerAccountUsingPOST","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ManagedUserVM"}}}},"responses":{"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/authorities":{"get":{"tags":["public-user-resource"],"summary":"getAuthorities","operationId":"getAuthoritiesUsingGET","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"type":"string"}}}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/users":{"get":{"tags":["public-user-resource"],"summary":"getAllPublicUsers","operationId":"getAllPublicUsersUsingGET","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserDTO"}}}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/admin/users":{"get":{"tags":["user-resource"],"summary":"getAllUsers","operationId":"getAllUsersUsingGET","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AdminUserDTO"}}}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"put":{"tags":["user-resource"],"summary":"updateUser","operationId":"updateUserUsingPUT","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdminUserDTO"}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AdminUserDTO"}}}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"post":{"tags":["user-resource"],"summary":"createUser","operationId":"createUserUsingPOST","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdminUserDTO"}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/admin/users/{login}":{"get":{"tags":["user-resource"],"summary":"getUser","operationId":"getUserUsingGET","parameters":[{"name":"login","in":"path","description":"login","required":true,"style":"simple","schema":{"pattern":"^(?>[a-zA-Z0-9!$&*+=?^_`{|}~.-]+@[a-zA-Z0-9-]+(?:\\.[a-zA-Z0-9-]+)*)|(?>[_.@A-Za-z0-9-]+)$","type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AdminUserDTO"}}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"delete":{"tags":["user-resource"],"summary":"deleteUser","operationId":"deleteUserUsingDELETE","parameters":[{"name":"login","in":"path","description":"login","required":true,"style":"simple","schema":{"pattern":"^(?>[a-zA-Z0-9!$&*+=?^_`{|}~.-]+@[a-zA-Z0-9-]+(?:\\.[a-zA-Z0-9-]+)*)|(?>[_.@A-Za-z0-9-]+)$","type":"string"}}],"responses":{"200":{"description":"OK"},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}}}},"components":{"schemas":{"AdminUserDTO":{"title":"AdminUserDTO","required":["login"],"type":"object","properties":{"activated":{"type":"boolean"},"authorities":{"uniqueItems":true,"type":"array","items":{"type":"string"}},"createdBy":{"type":"string"},"createdDate":{"type":"string","format":"date-time"},"email":{"maxLength":254,"minLength":5,"type":"string"},"firstName":{"maxLength":50,"minLength":0,"type":"string"},"id":{"type":"integer","format":"int64"},"imageUrl":{"maxLength":256,"minLength":0,"type":"string"},"langKey":{"maxLength":10,"minLength":2,"type":"string"},"lastModifiedBy":{"type":"string"},"lastModifiedDate":{"type":"string","format":"date-time"},"lastName":{"maxLength":50,"minLength":0,"type":"string"},"login":{"maxLength":50,"minLength":1,"pattern":"^(?>[a-zA-Z0-9!$&*+=?^_`{|}~.-]+@[a-zA-Z0-9-]+(?:\\.[a-zA-Z0-9-]+)*)|(?>[_.@A-Za-z0-9-]+)$","type":"string"}}},"KeyAndPasswordVM":{"title":"KeyAndPasswordVM","type":"object","properties":{"key":{"type":"string"},"newPassword":{"type":"string"}}},"ManagedUserVM":{"title":"ManagedUserVM","required":["login"],"type":"object","properties":{"activated":{"type":"boolean"},"authorities":{"uniqueItems":true,"type":"array","items":{"type":"string"}},"createdBy":{"type":"string"},"createdDate":{"type":"string","format":"date-time"},"email":{"maxLength":254,"minLength":5,"type":"string"},"firstName":{"maxLength":50,"minLength":0,"type":"string"},"id":{"type":"integer","format":"int64"},"imageUrl":{"maxLength":256,"minLength":0,"type":"string"},"langKey":{"maxLength":10,"minLength":2,"type":"string"},"lastModifiedBy":{"type":"string"},"lastModifiedDate":{"type":"string","format":"date-time"},"lastName":{"maxLength":50,"minLength":0,"type":"string"},"login":{"maxLength":50,"minLength":1,"pattern":"^(?>[a-zA-Z0-9!$&*+=?^_`{|}~.-]+@[a-zA-Z0-9-]+(?:\\.[a-zA-Z0-9-]+)*)|(?>[_.@A-Za-z0-9-]+)$","type":"string"},"password":{"maxLength":100,"minLength":4,"type":"string"}}},"PasswordChangeDTO":{"title":"PasswordChangeDTO","type":"object","properties":{"currentPassword":{"type":"string"},"newPassword":{"type":"string"}}},"PersistentToken":{"title":"PersistentToken","type":"object","properties":{"ipAddress":{"maxLength":39,"minLength":0,"type":"string"},"series":{"type":"string"},"tokenDate":{"type":"string","format":"date"},"userAgent":{"type":"string"}}},"User":{"title":"User","required":["activated","login"],"type":"object","properties":{"activated":{"type":"boolean"},"email":{"maxLength":254,"minLength":5,"type":"string"},"firstName":{"maxLength":50,"minLength":0,"type":"string"},"id":{"type":"integer","format":"int64"},"imageUrl":{"maxLength":256,"minLength":0,"type":"string"},"langKey":{"maxLength":10,"minLength":2,"type":"string"},"lastName":{"maxLength":50,"minLength":0,"type":"string"},"login":{"maxLength":50,"minLength":1,"pattern":"^(?>[a-zA-Z0-9!$&*+=?^_`{|}~.-]+@[a-zA-Z0-9-]+(?:\\.[a-zA-Z0-9-]+)*)|(?>[_.@A-Za-z0-9-]+)$","type":"string"},"resetDate":{"type":"string","format":"date-time"}}},"UserDTO":{"title":"UserDTO","type":"object","properties":{"id":{"type":"integer","format":"int64"},"login":{"type":"string"}}}}}}